Exract and fix whitelisted approval tests
